requirements_win.txt
python>=3.6
pyqt>=5
numpy
numexpr
scipy
h5py
gdal>=2.1
libgdal>=2.1
pillow
cython
matplotlib
lxml
scikit-image
pyreadline
mako
mkl-service